Initial Commit
Initial Commit

In this demo, I provide a procedural form of intelligence to NPCs, akin to animal pack behavior. The overarching idea is that NPCs will flee when they feel outnumbered and become bolder as they gain numbers until they decide they are strong enough to chase the player. Raycasts are used in conjunction with overlapshere for npc vision.